Added use cases excluding message-related ones
Changed GroupRepo to what I think is a fixed version (changed return to Group instead of List<Group> since there is no id for a list of groups) and added to UserRepo things to filter by when searching, although I'm not exactly sure what all possible filters are. For recommendGamer, I have just recommended based off of games so far since I believe that's what was agreed on.